packages/telegram-bot/requirements.txt
python-telegram-bot[webhooks]==21.6
python-dotenv>=1.0.0
faster-whisper==1.1.1
librosa==0.10.2
soundfile>=0.12.1
psycopg2-binary==2.9.9
requests==2.32.5
numpy==1.26.4
google-cloud-storage==2.18.2
scipy>=1.11.0
aiohttp>=3.9.0
